--- tests/peps/configs/config_U1_R_fermionic.py
@@ -1,7 +1,0 @@
-""" configuration of yastn tensor """
-import yastn.backend.backend_np as backend
-import yastn.sym.sym_U1 as sym
-default_dtype = 'float64'
-default_device = 'cpu'
-default_fusion = 'hard'
-
