--- tests/peps/configs/config_Z2_R_fermionic.py
@@ -1,7 +1,0 @@
-""" configuration of yastn tensor """
-import yastn.backend.backend_np as backend  # pylint: disable=unused-import
-import yastn.sym.sym_Z2 as sym  # pylint: disable=unused-import
-fermionic = True
-default_dtype = 'float64'
-default_device = 'cpu'
-default_fusion = 'hard'
